ci: lint（ruff）と型チェック（mypy）を導入（Issue #9）
## 設定（pyproject.toml）
- [tool.ruff]: line-length 110（日本語の全角文字ぶん少し広め）、target py311
- [tool.ruff.lint]: select = E,F,I,UP,B,C4,SIM,RUF。
  RUF001-003（ambiguous-unicode）は日本語コメント/文字列を誤検知するのでルールごと無効化。
- [tool.mypy]: files=["src"]、ignore_missing_imports、check_untyped_defs、
  warn_redundant_casts、warn_unused_ignores。
  numpy 同梱スタブ（3.12+ 構文）で落ちるので follow_imports=skip。
  view/tk_main_window.py は Tkinter アダプタで dict 展開の arg-type 誤検知が多いので
  そのファイルだけ arg-type を無効化。

## requirements-dev.txt
- ruff>=0.16, mypy>=1.11 を追加。

## CI（.github/workflows/test.yml）
- lint ジョブを追加（ubuntu 1 本）: `ruff check src tests` と `mypy`。

## 既存コードの指摘対応
- ruff --fix の安全な自動修正（import 並べ替え、Optional[X]→X|None、
  typing.Callable→collections.abc.Callable、未使用 noqa の削除 等、13 ファイル）。
- int(round(x)) → round(x)（round は ndigits 省略時 int を返す。7 箇所）。
- zip(...) の strict= 明示 / itertools.pairwise 化（frames.py・test_frames.py）。
- cli._make_reporter: dict の状態保持を nonlocal のクロージャ変数へ（mypy 型不整合）。
- config._ENV_MAP の型を Callable[[str], object] に、gui セクションは cast。
- pipeline.frames_index の注釈を Path | None → Path（実際に常に Path）。

## README
- 「既知の限界と次の一手」末尾の「CI・lint/型チェック・中間結果の再利用・LLM 落ち時の
  リトライは次フェーズで対応予定」は、CI（#8）・lint/型チェック（#9）導入済み、
  中間結果の再利用も「途中から再開」で実装済みなので、残る「LLM 落ち時のリトライ」
  だけに書き直し（ja/en）。

ローカル: ruff 全 pass / mypy 全 pass（21 ファイル）/ pytest 158 passed。

+++ src/meeting_minutes/model/vision.py
@@ -9,9 +9,9 @@ from __future__ import annotations
 import json
 import threading
 import time
+from collections.abc import Callable
 from dataclasses import asdict, dataclass
 from pathlib import Path
-from typing import Callable
 
 from .cancel import check_cancel
 from .config import load_prompt
@@ -38,7 +38,7 @@ class FrameNote:
 
 
 def _hhmmss(seconds: float) -> str:
-    seconds = max(0, int(round(seconds)))
+    seconds = max(0, round(seconds))
     h, rem = divmod(seconds, 3600)
     m, s = divmod(rem, 60)
     return f"{h:02d}:{m:02d}:{s:02d}"
@@ -55,7 +55,7 @@ def _format_elapsed(seconds: float) -> str:
     """処理にかかった時間の表示用（pipeline.py / minutes.py にも同名の複製がある）。"""
     if seconds < 60:
         return f"{seconds:.1f}秒"
-    minutes, sec = divmod(int(round(seconds)), 60)
+    minutes, sec = divmod(round(seconds), 60)
     if minutes < 60:
         return f"{minutes}分{sec:02d}秒"
     hours, minutes = divmod(minutes, 60)
@@ -133,7 +133,7 @@ def describe_frames(
         except LLMConnectionError:
             # 接続そのものが死んでいる場合は続けても無駄なので中断
             raise
-        except Exception as exc:  # noqa: BLE001 - 1 枚の失敗は握りつぶして続行
+        except Exception as exc:
             desc = f"(解析失敗: {exc})"
         elapsed = _format_elapsed(time.monotonic() - t0)
         notes.append(
